Refactor js code to ts
